myFinance/backend/portfolios/serializers.py
from rest_framework import serializers

from .models import Portfolio


class PortfolioSerializer(serializers.ModelSerializer):
    class Meta:
        model = Portfolio
        fields = ('id', 'name', 'created_at')
        read_only_fields = ('id', 'created_at')
