myFinance/backend/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
yfinance
pandas
numpy
scikit-learn
openai
